requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
altair==5.0.1
annotated-types==0.6.0
anyio==4.1.0
apache-airflow==2.8.1
apache-airflow-providers-common-io==1.3.0
apache-airflow-providers-common-sql==1.9.0
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-google==10.15.0
apache-airflow-providers-http==4.8.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-sqlite==3.6.0
apispec==6.3.0
argcomplete==3.2.1
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
backoff==2.2.1
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.1
cattrs==23.2.3
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.1.4
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.1.0
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.7
cycler==0.11.0
db-dtypes==1.2.0
debugpy==1.6.7.post1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.2
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2024.2.0
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.2.0
gcsfs==2024.2.0
gitdb==4.0.10
GitPython==3.1.32
google-ads==23.0.0
google-analytics-admin==0.22.5
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.42.1
google-cloud-appengine-logging==1.4.1
google-cloud-audit-log==0.2.5
google-cloud-automl==2.13.1
google-cloud-batch==0.17.11
google-cloud-bigquery==3.17.2
google-cloud-bigquery-datatransfer==3.14.1
google-cloud-bigquery-storage==2.24.0
google-cloud-bigtable==2.23.0
google-cloud-build==3.23.1
google-cloud-compute==1.16.1
google-cloud-container==2.40.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.18.1
google-cloud-dataflow-client==0.8.8
google-cloud-dataform==0.5.7
google-cloud-dataplex==1.12.1
google-cloud-dataproc==5.9.1
google-cloud-dataproc-metastore==1.15.1
google-cloud-dlp==3.15.1
google-cloud-kms==2.21.1
google-cloud-language==2.13.1
google-cloud-logging==3.9.0
google-cloud-memcache==1.9.1
google-cloud-monitoring==2.19.1
google-cloud-orchestration-airflow==1.11.1
google-cloud-os-login==2.14.1
google-cloud-pubsub==2.19.4
google-cloud-redis==2.15.1
google-cloud-resource-manager==1.12.1
google-cloud-run==0.10.3
google-cloud-secret-manager==2.18.1
google-cloud-spanner==3.42.0
google-cloud-speech==2.24.1
google-cloud-storage==2.14.0
google-cloud-storage-transfer==1.11.1
google-cloud-tasks==2.16.1
google-cloud-texttospeech==2.16.1
google-cloud-translate==3.15.1
google-cloud-videointelligence==2.13.1
google-cloud-vision==3.7.0
google-cloud-workflows==1.14.1
google-crc32c==1.5.0
google-re2==1.1
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
graphviz==0.20.1
greenlet==3.0.2
grpc-google-iam-v1==0.13.0
grpc-interceptor==0.15.4
grpcio==1.60.1
grpcio-gcp==0.2.2
grpcio-status==1.60.1
gspread==6.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
json-merge-patch==0.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
looker-sdk==24.0.0
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python==8.0.32
nest-asyncio==1.5.7
numpy==1.25.1
oauthlib==3.2.2
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
pandas-gbq==0.21.0
parso==0.8.3
pathspec==0.12.1
pendulum==2.1.2
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.9.1
plotly==5.15.0
pluggy==1.3.0
prison==0.2.1
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydata-google-auth==1.8.2
pydeck==0.8.0
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
pyOpenSSL==24.0.0
pyparsing==3.0.9
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.4.2
rich-argparse==1.4.0
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
setproctitle==1.3.3
shapely==2.0.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.50
sqlalchemy-bigquery==1.9.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.6.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.2
streamlit==1.25.0
StrEnum==0.4.15
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
unicodecsv==0.14.1
universal_pathlib==0.2.0
uritemplate==4.1.1
urllib3==2.0.4
validators==0.20.0
virtualenv==20.24.2
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.0.1
yarl==1.9.4
zipp==3.16.2
